+++ templates/temp-home.php
@@ -19,10 +19,10 @@
   <div class="content home">
 
     <? if (have_posts() ) : while (have_posts()) : the_post(); ?>
-      <section>
+      <article>
         <h1><? the_title(); ?></h1>
         <? the_content(); ?>
-      </section>
+      </article>
     <? endwhile; endif; ?>
 
   </div>